+++ service/src/index.ts
@@ -4,7 +4,7 @@ import 'dotenv/config'
 
 initDb()
 
-const port = Number(process.env.PORT) || 3001
+const port = Number(process.env.PORT) || 3508
 
 app.listen(port, () => {
   console.log(`🚀 zealous-admin-service running at http://localhost:${port}`)